+++ packages/gatsby/src/internal-plugins/query-runner/query-compiler.js
@@ -4,16 +4,17 @@ const normalize = require(`normalize-path`)
 import glob from "glob"
 
 import { validate } from "graphql"
-import { IRTransforms } from "relay-compiler"
-import RelayParser from "relay-compiler/lib/RelayParser"
-import ASTConvert from "relay-compiler/lib/ASTConvert"
-import GraphQLCompilerContext from "relay-compiler/lib/GraphQLCompilerContext"
-import filterContextForNode from "relay-compiler/lib/filterContextForNode"
+import { IRTransforms } from "@gatsbyjs/relay-compiler"
+import RelayParser from "@gatsbyjs/relay-compiler/lib/RelayParser"
+import ASTConvert from "@gatsbyjs/relay-compiler/lib/ASTConvert"
+import GraphQLCompilerContext from "@gatsbyjs/relay-compiler/lib/GraphQLCompilerContext"
+import filterContextForNode from "@gatsbyjs/relay-compiler/lib/filterContextForNode"
 const _ = require(`lodash`)
 
 import { store } from "../../redux"
+const { boundActionCreators } = require(`../../redux/actions`)
 import FileParser from "./file-parser"
-import GraphQLIRPrinter from "relay-compiler/lib/GraphQLIRPrinter"
+import GraphQLIRPrinter from "@gatsbyjs/relay-compiler/lib/GraphQLIRPrinter"
 import {
   graphqlError,
   graphqlValidationError,
@@ -62,11 +63,9 @@ const validationRules = [
 let lastRunHadErrors = null
 const overlayErrorID = `graphql-compiler`
 
-const resolveThemes = (plugins = []) =>
-  plugins.reduce((merged, plugin) => {
-    if (plugin.name.includes(`gatsby-theme-`)) {
-      merged.push(plugin.resolve)
-    }
+const resolveThemes = (themes = []) =>
+  themes.reduce((merged, theme) => {
+    merged.push(theme.themeDir)
     return merged
   }, [])
 
@@ -136,6 +135,7 @@ class Runner {
     const compiledNodes: Queries = new Map()
     const namePathMap = new Map()
     const nameDefMap = new Map()
+    const nameErrorMap = new Map()
     const documents = []
 
     for (let [filePath, doc] of nodes.entries()) {
@@ -143,6 +143,9 @@ class Runner {
 
       if (errors && errors.length) {
         this.reportError(graphqlValidationError(errors, filePath))
+        boundActionCreators.queryExtractionGraphQLError({
+          componentPath: filePath,
+        })
         return compiledNodes
       }
 
@@ -165,8 +168,18 @@ class Runner {
         )
       )
     } catch (error) {
-      this.reportError(graphqlError(namePathMap, nameDefMap, error))
-      return compiledNodes
+      const { formattedMessage, docName, message, codeBlock } = graphqlError(
+        namePathMap,
+        nameDefMap,
+        error
+      )
+      nameErrorMap.set(docName, { formattedMessage, message, codeBlock })
+      boundActionCreators.queryExtractionGraphQLError({
+        componentPath: namePathMap.get(docName),
+        error: formattedMessage,
+      })
+      this.reportError(formattedMessage)
+      return false
     }
 
     // relay-compiler v1.5.0 added "StripUnusedVariablesTransform" to
@@ -193,6 +206,9 @@ class Runner {
             otherNode && nameDefMap.get(otherNode.name)
           )
         )
+        boundActionCreators.queryExtractionGraphQLError({
+          componentPath: filePath,
+        })
         return
       }
 
@@ -206,6 +222,7 @@ class Runner {
         text,
         originalText: nameDefMap.get(name).text,
         path: filePath,
+        isHook: nameDefMap.get(name).isHook,
         isStaticQuery: nameDefMap.get(name).isStaticQuery,
         hash: nameDefMap.get(name).hash,
       }
@@ -217,6 +234,18 @@ class Runner {
             `${path.relative(store.getState().program.directory, filePath)}`
           )
       }
+
+      if (
+        query.isHook &&
+        process.env.NODE_ENV === `production` &&
+        typeof require(`react`).useContext !== `function`
+      ) {
+        report.panicOnBuild(
+          `You're likely using a version of React that doesn't support Hooks\n` +
+            `Please update React and ReactDOM to 16.8.0 or later to use the useStaticQuery hook.`
+        )
+      }
+
       compiledNodes.set(filePath, query)
     })
 
@@ -235,9 +264,13 @@ export { Runner, resolveThemes }
 
 export default async function compile(): Promise<Map<string, RootQuery>> {
   // TODO: swap plugins to themes
-  const { program, schema, plugins } = store.getState()
+  const { program, schema, themes } = store.getState()
 
-  const runner = new Runner(program.directory, resolveThemes(plugins), schema)
+  const runner = new Runner(
+    program.directory,
+    resolveThemes(themes.themes),
+    schema
+  )
 
   const queries = await runner.compileAll()