+++ www/src/utils/sidebar/create-link.js
@@ -1,10 +1,13 @@
 import React from "react"
 import { Link } from "gatsby"
 
-import presets, { colors } from "../presets"
+import { colors, space, transition, radii } from "../presets"
+import presets from "../../utils/sidebar/presets"
 
 const _getTitle = (title, isDraft) => (isDraft ? title.slice(0, -1) : title)
 const _isDraft = title => title.slice(-1) === `*`
+
+const bulletSize = 8
 
 const createLink = ({
   item,
@@ -13,6 +16,8 @@ const createLink = ({
   isParentOfActiveItem,
   stepsUI,
   customCSS,
+  level,
+  indention,
 }) => {
   const isDraft = _isDraft(item.title)
   const title = _getTitle(item.title, isDraft)
@@ -24,14 +29,14 @@ const createLink = ({
         alignItems: `center`,
         position: `relative`,
         "&:before": {
-          background: colors.ui.border,
+          background: presets.itemBorderColor,
           bottom: 0,
           top: `auto`,
           content: `''`,
           height: 1,
           position: `absolute`,
           right: 0,
-          left: 0,
+          left: indention,
         },
       }}
     >
@@ -42,6 +47,41 @@ const createLink = ({
           isActive && styles.activeLink,
           isParentOfActiveItem && styles.parentOfActiveLink,
           customCSS && customCSS,
+          { paddingLeft: indention },
+          {
+            "&:before, &:after": {
+              left:
+                level === 0 || level === 1
+                  ? `calc(${indention} - 16px)`
+                  : `calc(${indention} - 24px)`,
+              top: `1.3em`,
+              height: bulletSize,
+              position: `absolute`,
+              transition: `all ${transition.speed.default} ${
+                transition.curve.default
+              }`,
+            },
+
+            "&:before": {
+              background: isActive ? colors.gatsby : false,
+              borderRadius: radii[6],
+              content: `''`,
+              transform: isActive ? `scale(1)` : `scale(0.1)`,
+              width: bulletSize,
+            },
+            "&:after": {
+              background: colors.gatsby,
+              borderRadius: radii[2],
+              content: `''`,
+              left:
+                level === 0 || level === 1
+                  ? `calc(${indention} - 8px)`
+                  : `calc(${indention} - 16px)`,
+              opacity: isActive ? 1 : 0,
+              transform: `translateX(-100px)`,
+              width: isActive ? 100 : 1,
+            },
+          },
         ]}
         onClick={onLinkClick}
         to={item.link}
@@ -53,18 +93,6 @@ const createLink = ({
   )
 }
 
-const bulletOffset = {
-  default: {
-    left: -25,
-    top: `1.15em`,
-  },
-  desktop: {
-    top: `1.2em`,
-  },
-}
-
-const bulletSize = 8
-
 const styles = {
   draft: {
     "&&": {
@@ -74,37 +102,30 @@ const styles = {
   parentOfActiveLink: {
     "&&": {
       color: colors.gatsby,
-      fontWeight: `bold`,
+      fontWeight: 600,
     },
   },
   activeLink: {
     "&&": {
       color: colors.gatsby,
-      fontWeight: `bold`,
-    },
-    "&:before": {
-      background: colors.gatsby,
-      transform: `scale(1)`,
-    },
-    "&:after": {
-      width: 200,
-      opacity: 1,
+      fontWeight: 600,
+      background: presets.activeSectionBackground,
     },
   },
   link: {
-    paddingRight: 40,
-    minHeight: 40,
-    paddingTop: 10,
-    paddingBottom: 10,
+    paddingRight: space[4],
+    minHeight: presets.itemMinHeight,
+    paddingTop: space[3],
+    paddingBottom: space[3],
     position: `relative`,
     zIndex: 1,
     width: `100%`,
     "&&": {
       border: 0,
-      boxShadow: `none`,
+      color: colors.gray.copy,
       fontWeight: `normal`,
       "&:hover": {
-        background: `transparent`,
+        background: presets.itemHoverBackground,
         color: colors.gatsby,
         "&:before": {
           background: colors.gatsby,
@@ -112,50 +133,19 @@ const styles = {
         },
       },
     },
-    "&:before, &:after": {
-      ...bulletOffset.default,
-      height: bulletSize,
-      position: `absolute`,
-      transition: `all ${presets.animation.speedDefault} ${
-        presets.animation.curveDefault
-      }`,
-    },
-    "&:before": {
-      borderRadius: `100%`,
-      content: `''`,
-      transform: `scale(0.1)`,
-      width: bulletSize,
-      [presets.Tablet]: {
-        ...bulletOffset.desktop,
-      },
-    },
-    "&:after": {
-      background: colors.gatsby,
-      borderRadius: 4,
-      content: `''`,
-      left: bulletOffset.default.left + 7,
-      opacity: 0,
-      transform: `translateX(-200px)`,
-      width: 1,
-      [presets.Tablet]: {
-        ...bulletOffset.desktop,
-      },
-    },
   },
   subsectionLink: {
-    ...bulletOffset.default,
-    background: `#fff`,
+    left: space[6],
+    background: colors.white,
     border: `1px solid ${colors.ui.bright}`,
-    borderRadius: `100%`,
+    borderRadius: radii[6],
     display: `block`,
     fontWeight: `normal`,
     height: bulletSize,
     position: `absolute`,
     width: bulletSize,
+    top: `1.3em`,
     zIndex: -1,
-    [presets.Tablet]: {
-      ...bulletOffset.desktop,
-    },
   },
 }